rfc: rename index to scene and rename ui to index
